Fix simulation uptime
r7110
2022-08-07

1. https://github.com/mars-sim/mars-sim/issues/655

1. Revise updateTime() to use the standard time as the param.

+++ mars-sim-core/src/main/java/org/mars_sim/msp/core/time/UpTimer.java
@@ -53,9 +53,12 @@ public class UpTimer implements Serializable {
     }
 
     /**
-     * TODO: Why have this method? Just use the System millisecodns but add a substtraction for the time it is paused.
+     * Updates the simulation uptime.
+     * TODO: May switch to using System.currentTimeMillis() to get the start time and end time and subtract the pause time.
+     * 
+     * @param elapsedMilli
      */
-    public void updateTime(long elapsedMilli) {
+    public void updateTime(double elapsedMilli) {
         uptime += elapsedMilli;
     }
